+++ src/main/java/trabajo/Recepcionista.java
@@ -16,7 +16,7 @@ public class Recepcionista extends Empleado{
     
     // polimorfismo
     @Override
-    public double sueldo(int hrs){
+    public int sueldo(int hrs){
         int extra = 0;
         int hrsext = hrs - 48;
         
@@ -28,7 +28,7 @@ public class Recepcionista extends Empleado{
             extra = hrsext * 300;
         }
         
-        double s = hrs + extra;
+        int s = hrs + extra;
         
         return s;
     }
